% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{fivestar}
\alias{fivestar}
\title{Add a star}
\usage{
fivestar(
  rating,
  text = round(rating, 1),
  aria_label = paste("The rating out of five stars is", text),
  star_size = NULL,
  star_color = NULL,
  star_background = NULL
)
}
\arguments{
\item{rating}{rating out of five}

\item{text}{text to display in tag, defaults to the rating rounded to one decimal place}

\item{aria_label}{accessibility label for text, defaults to "rating}

\item{star_size}{CSS font-size for star background, default is to inherit from \code{use_fivestars()}}

\item{star_color}{CSS colour for stars, default is to inherit from \code{use_fivestars()}}

\item{star_background}{CSS colour for star background, default is to inherit from \code{use_fivestars()}}
}
\value{
an html tag that renders a star rating out of five
}
\description{
This function creates an HTML tag with the appropriate number of stars, as well as an accessibility label.
}
\examples{
fivestar(3.5)

}
