% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{fivestars-package}
\alias{fivestars}
\alias{fivestars-package}
\title{fivestars: Generates HTML/CSS star ratings}
\description{
Creates HTML and CSS star ratings in web documents.
}
\author{
\strong{Maintainer}: Tan Ho \email{tan@tanho.ca} (\href{https://orcid.org/0000-0001-8388-5155}{ORCID})

}
\keyword{internal}
