% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{fivestars}
\alias{fivestars}
\title{Add a star}
\usage{
fivestars(
  rating,
  label = round(rating, 1),
  aria_label = paste("The rating is", round(rating, 1), "out of five stars"),
  label_pos = "left",
  star_size = NULL,
  star_color = NULL,
  star_background = NULL
)
}
\arguments{
\item{rating}{rating out of five}

\item{label}{text to display in tag, defaults to the rating rounded to one decimal place}

\item{aria_label}{accessibility label for text}

\item{label_pos}{"left" or "right" - position of label relative to stars}

\item{star_size}{CSS font-size for star background, default is to inherit from \code{use_fivestars()}}

\item{star_color}{CSS colour for stars, default is to inherit from \code{use_fivestars()}}

\item{star_background}{CSS colour for star background, default is to inherit from \code{use_fivestars()}}
}
\value{
an html tag that renders a star rating out of five
}
\description{
This function creates an HTML tag with the appropriate number of stars, as well as an accessibility label.
}
\examples{
use_fivestars()
fivestars(3.5)

}
