% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{use_fivestars}
\alias{use_fivestars}
\title{Add stars CSS}
\usage{
use_fivestars(
  star_size = "larger",
  star_color = "#fc0",
  star_background = "#999"
)
}
\arguments{
\item{star_size}{CSS font size for your stars, default is \code{"larger"}}

\item{star_color}{CSS colour for star, default is \verb{#fc0}}

\item{star_background}{CSS colour for star background, default is \verb{#999}}
}
\value{
HTML that inserts CSS into the page head
}
\description{
This function adds the appropriate CSS to your page for star ratings.
}
\examples{
use_fivestars()

}
